Major Update
Changelog :
- Args from cryptoprice are now free from uppercase
- Tried to make a real help command
- Started to assemble serverinfo and hoststats
- Created a logout and restart command for client
- Added permission to say command
- Transformed command name from file to name stored in file
- Failed to log the console again

Upcoming :
- Real help command
- Improved info command

+++ command/say.js
@@ -1,3 +1,5 @@
+const { admin } = require('../settings.json')
+
 module.exports.help = {
     name : "say",
     description: 'Say command',
@@ -6,6 +8,8 @@ module.exports.help = {
 };
 
 module.exports.execute = async (client, message, args) => {
-    message.channel.bulkDelete(1)
-    message.channel.send(args.join(' '))
+    if (message.author.id == admin || message.member.hasPermission('ADMINISTRATOR')) {
+        message.channel.bulkDelete(1)
+        message.channel.send(args.join(' '))
+    }
 };
